GEODE-3286 Failing to cleanup connections from ConnectionTable
In previous fix, we were checking "if thread is stopped then add
connection to receiver list". Instead, it should be if thread is stopped
then we should not add connection to receiver list. As a fix, we add
connection to reciver list if connection is not closed or thread is not
stopped.

Added unit for it.